put checkers in a separate file from base class
