+++ snuba/datasets/storages/errors.py
@@ -14,7 +14,7 @@ from snuba.datasets.storages.errors_common import (
 )
 from snuba.datasets.storages.storage_key import StorageKey
 from snuba.datasets.table_storage import build_kafka_stream_loader_from_settings
-from snuba.query.processors.condition_checkers import ProjectIdEnforcer
+from snuba.query.processors.condition_checkers.checkers import ProjectIdEnforcer
 from snuba.replacers.replacer_processor import ReplacerState
 from snuba.subscriptions.utils import SchedulingWatermarkMode
 from snuba.utils.streams.topics import Topic
