+++ snuba/datasets/storages/profiles.py
@@ -7,7 +7,10 @@ from snuba.datasets.schemas.tables import WritableTableSchema
 from snuba.datasets.storage import WritableTableStorage
 from snuba.datasets.storages.storage_key import StorageKey
 from snuba.datasets.table_storage import build_kafka_stream_loader_from_settings
-from snuba.query.processors.condition_checkers import OrgIdEnforcer, ProjectIdEnforcer
+from snuba.query.processors.condition_checkers.checkers import (
+    OrgIdEnforcer,
+    ProjectIdEnforcer,
+)
 from snuba.query.processors.physical.table_rate_limit import TableRateLimit
 from snuba.query.processors.physical.type_converters.uuid_column_processor import (
     UUIDColumnProcessor,
