+++ snuba/datasets/storages/replays.py
@@ -14,7 +14,7 @@ from snuba.datasets.schemas.tables import WritableTableSchema
 from snuba.datasets.storage import WritableTableStorage
 from snuba.datasets.storages.storage_key import StorageKey
 from snuba.datasets.table_storage import build_kafka_stream_loader_from_settings
-from snuba.query.processors.condition_checkers import ProjectIdEnforcer
+from snuba.query.processors.condition_checkers.checkers import ProjectIdEnforcer
 from snuba.query.processors.physical.table_rate_limit import TableRateLimit
 from snuba.utils.schemas import Nested
 from snuba.utils.streams.configuration_builder import build_kafka_producer_configuration
