+++ snuba/datasets/storages/sessions.py
@@ -17,7 +17,10 @@ from snuba.datasets.storage import ReadableTableStorage, WritableTableStorage
 from snuba.datasets.storages.storage_key import StorageKey
 from snuba.datasets.table_storage import build_kafka_stream_loader_from_settings
 from snuba.query.exceptions import ValidationException
-from snuba.query.processors.condition_checkers import OrgIdEnforcer, ProjectIdEnforcer
+from snuba.query.processors.condition_checkers.checkers import (
+    OrgIdEnforcer,
+    ProjectIdEnforcer,
+)
 from snuba.query.processors.physical import ClickhouseQueryProcessor
 from snuba.query.processors.physical.prewhere import PrewhereProcessor
 from snuba.query.processors.physical.table_rate_limit import TableRateLimit
