+++ snuba/datasets/storages/transactions_common.py
@@ -11,7 +11,7 @@ from snuba.clickhouse.columns import (
 from snuba.clickhouse.columns import SchemaModifiers as Modifiers
 from snuba.clickhouse.columns import String, UInt
 from snuba.datasets.plans.splitters.strategies import TimeSplitQueryStrategy
-from snuba.query.processors.condition_checkers import ProjectIdEnforcer
+from snuba.query.processors.condition_checkers.checkers import ProjectIdEnforcer
 from snuba.query.processors.physical.array_has_optimizer import ArrayHasOptimizer
 from snuba.query.processors.physical.arrayjoin_keyvalue_optimizer import (
     ArrayJoinKeyValueOptimizer,
